+++ src/lib.rs
@@ -3,6 +3,7 @@ pub mod config;
 pub mod event;
 pub mod index;
 pub mod link;
+pub mod markdown;
 pub mod note;
 pub mod tree;
 pub mod ui;